- Added Try&Catch for I/O Operations, and logging

+++ Server/Data.cs
@@ -2,43 +2,16 @@
 
 class Data
 {
+    private readonly ILogger _logger;
     private List<Recipe> _recipes { get; set; } = new();
     private List<string> _categories { get; set; } = new();
     private string _recipesFilePath;
     private string _categoriesFilePath;
-
-    public Data()
+    public Data(ILogger logger)
     {
         _recipesFilePath = Path.Combine(Environment.CurrentDirectory, "Data", "Recipes.json");
         _categoriesFilePath = Path.Combine(Environment.CurrentDirectory, "Data", "Categories.json");
-    }
-
-    public async Task LoadData()
-    {
-        if (!File.Exists(_recipesFilePath))
-        {
-            _recipes = new List<Recipe>();
-            await File.WriteAllTextAsync(_recipesFilePath, JsonSerializer.Serialize(_recipes));
-        }
-        if (!File.Exists(this._categoriesFilePath))
-        {
-            _recipes = new List<Recipe>();
-            await File.WriteAllTextAsync(_categoriesFilePath, JsonSerializer.Serialize(_recipes));
-        }
-        using (StreamReader r = new(_recipesFilePath))
-        {
-            var data = await r.ReadToEndAsync();
-            var json = JsonSerializer.Deserialize<List<Recipe>>(data);
-            if (json != null)
-                _recipes = json;
-        }
-        using (StreamReader r = new(_categoriesFilePath))
-        {
-            var data = await r.ReadToEndAsync();
-            var json = JsonSerializer.Deserialize<List<string>>(data);
-            if (json != null)
-                _categories = json;
-        }
+        _logger = logger;
     }
 
     public async Task<List<Recipe>> GetRecipesAsync()
@@ -66,7 +39,7 @@ class Data
     public async Task RemoveRecipeAsync(Guid id)
     {
         await LoadData();
-        var recipe = await GetRecipeAsync(id); 
+        var recipe = await GetRecipeAsync(id);
         _recipes.Remove(recipe);
         await SaveDataAsync();
     }
@@ -106,7 +79,7 @@ class Data
         recipe.Instructions = newInstructions;
         await SaveDataAsync();
     }
-    
+
     public async Task EditCategoryAsync(Guid id, string category, string newCategory)
     {
         await LoadData();
@@ -125,7 +98,7 @@ class Data
     public async Task AddCategoryAsync(string category)
     {
         await LoadData();
-        if(!_categories.Contains(category))
+        if (!_categories.Contains(category))
             _categories.Add(category);
         await SaveDataAsync();
     }
@@ -135,7 +108,7 @@ class Data
         await LoadData();
         if (_categories.Contains(category))
         {
-            foreach(var recipe in _recipes.Where(r => r.Categories.Contains(category)))
+            foreach (var recipe in _recipes.Where(r => r.Categories.Contains(category)))
             {
                 recipe.Categories.Remove(category);
                 recipe.Categories.Add(newCategory);
@@ -152,7 +125,7 @@ class Data
         if (_categories.Contains(category))
         {
             _categories.Remove(category);
-            var recipes =_recipes.Where(r => r.Categories.Contains(category)).ToList();
+            var recipes = _recipes.Where(r => r.Categories.Contains(category)).ToList();
             foreach (var recipe in recipes)
                 recipe.Categories.Remove(category);
         }
@@ -177,9 +150,60 @@ class Data
         await SaveDataAsync();
     }
 
+    public async Task LoadData()
+    {
+        // Creating the files if they do not exist
+        try
+        {
+            if (!File.Exists(_recipesFilePath))
+            {
+                _recipes = new List<Recipe>();
+                await File.WriteAllTextAsync(_recipesFilePath, JsonSerializer.Serialize(_recipes));
+            }
+            if (!File.Exists(this._categoriesFilePath))
+            {
+                _recipes = new List<Recipe>();
+                await File.WriteAllTextAsync(_categoriesFilePath, JsonSerializer.Serialize(_recipes));
+            }
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(DateTime.Now.ToString(), "- Error while attempting to create the files: ", ex.ToString());
+        }
+        try
+        {
+            using (StreamReader r = new(_recipesFilePath))
+            {
+                var data = await r.ReadToEndAsync();
+                var json = JsonSerializer.Deserialize<List<Recipe>>(data);
+                if (json != null)
+                    _recipes = json;
+            }
+            using (StreamReader r = new(_categoriesFilePath))
+            {
+                var data = await r.ReadToEndAsync();
+                var json = JsonSerializer.Deserialize<List<string>>(data);
+                if (json != null)
+                    _categories = json;
+            }
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(DateTime.Now.ToString(), "- Error while attempting to save the data to the files: ", ex.ToString());
+        }
+
+    }
+
     public async Task SaveDataAsync()
     {
-        await File.WriteAllTextAsync(_recipesFilePath, JsonSerializer.Serialize(_recipes));
-        await File.WriteAllTextAsync(_categoriesFilePath, JsonSerializer.Serialize(_categories));
+        try
+        {
+            await File.WriteAllTextAsync(_recipesFilePath, JsonSerializer.Serialize(_recipes));
+            await File.WriteAllTextAsync(_categoriesFilePath, JsonSerializer.Serialize(_categories));
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(DateTime.Now.ToString(), "- Error while attempting to save files: ", ex.ToString());
+        }
     }
 }
